Move local_lagrange header files to include/
