+++ src/lib/local_lagrange/local_lagrange.cpp
@@ -1,3 +1,1 @@
-#include "local_lagrange.h"
-
-
+#include <local_lagrange/local_lagrange.h>
