+++ src/lib/local_lagrange/local_lagrange_assembler.cpp
@@ -1,1 +1,1 @@
-#include "local_lagrange_assembler.h"
+#include <local_lagrange/local_lagrange_assembler.h>
