+++ src/lib/local_lagrange/local_lagrange_interpolant.cpp
@@ -1,2 +1,1 @@
-#include "local_lagrange_interpolant.h"
-
+#include <local_lagrange/local_lagrange_interpolant.h>
