+++ src/lib/local_lagrange/test/integration_test.cpp
@@ -1,5 +1,5 @@
-#include "../local_lagrange.h"
-#include "../local_lagrange_assembler.h"
+#include <local_lagrange/local_lagrange.h>
+#include <local_lagrange/local_lagrange_assembler.h>
 
 #include <boost/geometry.hpp>
 #include <boost/geometry/geometries/box.hpp>
