+++ src/lib/local_lagrange/test/speed_test.cpp
@@ -1,4 +1,4 @@
-#include "../local_lagrange_interpolant.h"
+#include <local_lagrange/local_lagrange_interpolant.h>
 
 #include <math_utils/math_tools.h>
 #include <gtest/gtest.h>
